chore(core): moved all gatekeeper functions to a single lib file
