+++ engine/lib/group.php
@@ -5,42 +5,6 @@
  * mark any given container as their container.
  */
 
-use Elgg\Exceptions\Http\Gatekeeper\GroupToolGatekeeperException;
-
-/**
- * Checks if a group has a specific tool enabled.
- * Forward to the group if the tool is disabled.
- *
- * @param string $option     The group tool option to check
- * @param int    $group_guid The group that owns the page. If not set, this
- *                           will be pulled from elgg_get_page_owner_guid().
- *
- * @return void
- * @throws GroupToolGatekeeperException
- * @since 3.0.0
- */
-function elgg_group_tool_gatekeeper($option, $group_guid = null) {
-	$group_guid = $group_guid ?: elgg_get_page_owner_guid();
-	
-	$group = get_entity($group_guid);
-	if (!$group instanceof \ElggGroup) {
-		return;
-	}
-	
-	if ($group->isToolEnabled($option)) {
-		return;
-	}
-
-	$ex = new GroupToolGatekeeperException();
-	$ex->setRedirectUrl($group->getURL());
-	$ex->setParams([
-		'entity' => $group,
-		'tool' => $option,
-	]);
-
-	throw $ex;
-}
-
 /**
  * Allow group members to write to the group container
  *
