+++ engine/lib/pagehandler.php
@@ -125,65 +125,6 @@ function elgg_generate_action_url($action, array $query = [], $add_csrf_tokens =
 }
 
 /**
- * Used at the top of a page to mark it as logged in users only.
- *
- * @return void
- * @throws \Elgg\Exceptions\Http\Gatekeeper\LoggedInGatekeeperException
- * @since 1.9.0
- */
-function elgg_gatekeeper() {
-	_elgg_services()->gatekeeper->assertAuthenticatedUser();
-}
-
-/**
- * Used at the top of a page to mark it as admin only.
- *
- * @return void
- * @throws \Elgg\Exceptions\Http\Gatekeeper\AdminGatekeeperException
- * @since 1.9.0
- */
-function elgg_admin_gatekeeper() {
-	_elgg_services()->gatekeeper->assertAuthenticatedAdmin();
-}
-
-/**
- * Can the viewer see this entity?
- *
- * Tests if the entity exists and whether the viewer has access to the entity
- * if it does. If the viewer cannot view this entity, it forwards to an
- * appropriate page.
- *
- * @param int    $guid              Entity GUID
- * @param string $type              Optional required entity type
- * @param string $subtype           Optional required entity subtype
- * @param bool   $validate_can_edit flag to check canEdit access
- *
- * @return void
- *
- * @throws Exception
- * @throws \Elgg\Exceptions\Http\EntityNotFoundException
- * @throws \Elgg\Exceptions\Http\EntityPermissionsException
- * @throws \Elgg\Exceptions\HttpException
- * @since 1.9.0
- */
-function elgg_entity_gatekeeper($guid, $type = null, $subtype = null, $validate_can_edit = false) {
-	$entity = _elgg_services()->gatekeeper->assertExists($guid, $type, $subtype);
-	_elgg_services()->gatekeeper->assertAccessibleEntity($entity, null, $validate_can_edit);
-}
-
-/**
- * Require that the current request be an XHR. If not, execution of the current function
- * will end and a 400 response page will be sent.
- *
- * @return void
- * @throws \Elgg\Exceptions\Http\Gatekeeper\AjaxGatekeeperException
- * @since 1.12.0
- */
-function elgg_ajax_gatekeeper() {
-	_elgg_services()->gatekeeper->assertXmlHttpRequest();
-}
-
-/**
  * Prepares a successful response to be returned by a page or an action handler
  *
  * @param mixed  $content     Response content
